Add ThemeProvider, update layout and styles, and enhance dependencies
