+++ lib/fonts.ts
@@ -1,0 +1,13 @@
+import { Geist, Geist_Mono } from "next/font/google";
+
+const geistSans = Geist({
+	variable: "--font-geist-sans",
+	subsets: ["latin"],
+});
+
+const geistMono = Geist_Mono({
+	variable: "--font-geist-mono",
+	subsets: ["latin"],
+});
+
+export { geistSans, geistMono };
